requirements.txt
# Core dependencies
anthropic>=0.18.0
openai>=1.12.0
google-generativeai>=0.3.0
pydantic>=2.0.0
pyyaml>=6.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Utilities
python-dotenv>=1.0.0
loguru>=0.7.0

# Optional: Para futura integración de modelos locales
# torch>=2.0.0
# transformers>=4.30.0